requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-dotenv==1.0.1
google-generativeai==0.3.2
python-multipart==0.0.7
jinja2==3.1.3
starlette==0.36.3